+++ app/src/reducers/report.js
@@ -26,14 +26,12 @@ export default function (state = initialState, action) {
       const polygonData = action.payload.polygonData;
       const id = polygonData.cartodb_id;
 
-      let reportingId = polygonData.cartodb_id;
-      if (polygonData.reporting_id !== undefined) {
-        reportingId = polygonData.reporting_id;
-      } else if (polygonData.reportingId !== undefined) {
-        reportingId = polygonData.reportingId;
-      }
+      const reportingId = [polygonData.reporting_id, polygonData.report_id, polygonData.reportingId, polygonData.cartodb_id]
+        .find(e => e !== undefined);
 
-      const name = (polygonData.name !== undefined) ? polygonData.name : polygonData.cartodb_id.toString();
+      const name = [polygonData.reporting_name, polygonData.name, polygonData.cartodb_id]
+        .find(e => e !== undefined);
+
       const isInReport = !!state.polygons.find(polygon => polygon.id === id);
       return Object.assign({}, state, {
         currentPolygon: {
@@ -80,10 +78,12 @@ export default function (state = initialState, action) {
         layerId: action.payload.layerId,
         reportId: action.payload.reportId,
         layerTitle: action.payload.layerTitle,
-        status: REPORT_STATUS.idle
+        status: REPORT_STATUS.idle,
+        polygonsIds: [],
+        currentPolygon: {}
       });
     case DISCARD_REPORT:
-      return Object.assign({}, state, { polygons: [], layerId: null });
+      return Object.assign({}, state, { polygons: [], polygonsIds: [], currentPolygon: {}, layerId: null });
 
     case SET_REPORT_STATUS_SENT:
       return Object.assign({}, state, { status: REPORT_STATUS.sent, statusText: action.payload });